Start factoring out some common parsing infrastructure.
